removing pyi file
