+++ src/tracksdata/attrs.py
@@ -2,10 +2,10 @@ import functools
 import math
 import operator
 from collections.abc import Callable, Sequence
-from typing import Any, Union
+from typing import Any, Union, overload
 
-import polars as pl
 import numpy as np
+import polars as pl
 from polars import DataFrame, Expr, Series
 
 Scalar = int | float | str | bool | complex | np.number
@@ -59,6 +59,42 @@ class AttrComparison:
 
     def _delegate_operator(self, other: ExprInput, op: Callable[[Expr, Expr], Expr], reverse: bool = False) -> "Attr":
         return self.to_attr()._delegate_operator(other, op, reverse)
+
+    # Binary operators
+    def __add__(self, other: ExprInput) -> "Attr": ...
+    def __sub__(self, other: ExprInput) -> "Attr": ...
+    def __mul__(self, other: ExprInput) -> "Attr": ...
+    def __truediv__(self, other: ExprInput) -> "Attr": ...
+    def __floordiv__(self, other: ExprInput) -> "Attr": ...
+    def __mod__(self, other: ExprInput) -> "Attr": ...
+    def __pow__(self, other: ExprInput) -> "Attr": ...
+    def __and__(self, other: ExprInput) -> "Attr": ...
+    def __or__(self, other: ExprInput) -> "Attr": ...
+    def __xor__(self, other: ExprInput) -> "Attr": ...
+
+    # Reverse operators
+    def __radd__(self, other: Scalar) -> "Attr": ...
+    def __rsub__(self, other: Scalar) -> "Attr": ...
+    def __rmul__(self, other: Scalar) -> "Attr": ...
+    def __rtruediv__(self, other: Scalar) -> "Attr": ...
+    def __rfloordiv__(self, other: Scalar) -> "Attr": ...
+    def __rmod__(self, other: Scalar) -> "Attr": ...
+    def __rpow__(self, other: Scalar) -> "Attr": ...
+    def __rand__(self, other: Scalar) -> "Attr": ...
+    def __ror__(self, other: Scalar) -> "Attr": ...
+    def __rxor__(self, other: Scalar) -> "Attr": ...
+
+    # Comparison operators (always return Attr)
+    @overload
+    def __eq__(self, other: "Attr") -> "Attr": ...
+    @overload
+    def __eq__(self, other: Scalar) -> "AttrComparison": ...
+    def __eq__(self, other): ...
+    def __ne__(self, other: ExprInput) -> "Attr": ...
+    def __lt__(self, other: ExprInput) -> "Attr": ...
+    def __le__(self, other: ExprInput) -> "Attr": ...
+    def __gt__(self, other: ExprInput) -> "Attr": ...
+    def __ge__(self, other: ExprInput) -> "Attr": ...
 
 
 class Attr:
@@ -165,11 +201,11 @@ class Attr:
             return self._delegate_operator(other, op, reverse)
 
         if reverse:
-            lhs = other
+            lhs = Attr(other)
             rhs = self
         else:
             lhs = self
-            rhs = other
+            rhs = Attr(other)
         return AttrComparison(lhs, op, rhs)
 
     def alias(self, name: str) -> "Attr":
@@ -269,6 +305,67 @@ class Attr:
 
     def __repr__(self) -> str:
         return f"AttrExpr({self.expr})"
+
+    # Binary operators
+    def __add__(self, other: ExprInput) -> "Attr": ...
+    def __sub__(self, other: ExprInput) -> "Attr": ...
+    def __mul__(self, other: ExprInput) -> "Attr": ...
+    def __truediv__(self, other: ExprInput) -> "Attr": ...
+    def __floordiv__(self, other: ExprInput) -> "Attr": ...
+    def __mod__(self, other: ExprInput) -> "Attr": ...
+    def __pow__(self, other: ExprInput) -> "Attr": ...
+    def __and__(self, other: ExprInput) -> "Attr": ...
+    def __or__(self, other: ExprInput) -> "Attr": ...
+    def __xor__(self, other: ExprInput) -> "Attr": ...
+
+    # Reverse operators
+    def __radd__(self, other: Scalar) -> "Attr": ...
+    def __rsub__(self, other: Scalar) -> "Attr": ...
+    def __rmul__(self, other: Scalar) -> "Attr": ...
+    def __rtruediv__(self, other: Scalar) -> "Attr": ...
+    def __rfloordiv__(self, other: Scalar) -> "Attr": ...
+    def __rmod__(self, other: Scalar) -> "Attr": ...
+    def __rpow__(self, other: Scalar) -> "Attr": ...
+    def __rand__(self, other: Scalar) -> "Attr": ...
+    def __ror__(self, other: Scalar) -> "Attr": ...
+    def __rxor__(self, other: Scalar) -> "Attr": ...
+
+    # Comparison operators with overloads
+    @overload
+    def __eq__(self, other: "Attr") -> "Attr": ...
+    @overload
+    def __eq__(self, other: Scalar) -> "AttrComparison": ...
+    def __eq__(self, other): ...
+
+    @overload
+    def __ne__(self, other: "Attr") -> "Attr": ...
+    @overload
+    def __ne__(self, other: Scalar) -> "AttrComparison": ...
+    def __ne__(self, other): ...
+
+    @overload
+    def __lt__(self, other: "Attr") -> "Attr": ...
+    @overload
+    def __lt__(self, other: Scalar) -> "AttrComparison": ...
+    def __lt__(self, other): ...
+
+    @overload
+    def __le__(self, other: "Attr") -> "Attr": ...
+    @overload
+    def __le__(self, other: Scalar) -> "AttrComparison": ...
+    def __le__(self, other): ...
+
+    @overload
+    def __gt__(self, other: "Attr") -> "Attr": ...
+    @overload
+    def __gt__(self, other: Scalar) -> "AttrComparison": ...
+    def __gt__(self, other): ...
+
+    @overload
+    def __ge__(self, other: "Attr") -> "Attr": ...
+    @overload
+    def __ge__(self, other: Scalar) -> "AttrComparison": ...
+    def __ge__(self, other): ...
 
 
 # Auto-generate operator methods using functools.partialmethod
